+++ src/pyproject_hooks/_impl.py
@@ -8,17 +8,7 @@ from os.path import abspath
 from os.path import join as pjoin
 from subprocess import STDOUT, check_call, check_output
 
-from .in_process import _in_proc_script_path
-
-__all__ = [
-    'BackendUnavailable',
-    'BackendInvalid',
-    'HookMissing',
-    'UnsupportedOperation',
-    'default_subprocess_runner',
-    'quiet_subprocess_runner',
-    'Pep517HookCaller',
-]
+from ._in_process import _in_proc_script_path
 
 
 def write_json(obj, path, **kwargs):
@@ -100,7 +90,7 @@ def norm_and_check(source_tree, requested):
     return abs_requested
 
 
-class Pep517HookCaller:
+class BuildBackendHookCaller:
     """A wrapper around a source directory to be built with a PEP 517 backend.
 
     :param source_dir: The path to the source directory, containing