+++ tests/test_call_hooks.py
@@ -10,10 +10,10 @@ import pytest
 from testpath import assert_isfile, modified_env
 from testpath.tempdir import TemporaryDirectory, TemporaryWorkingDirectory
 
-from pep517._compat import tomllib
-from pep517.wrappers import (
+from pyproject_hooks._compat import tomllib
+from pyproject_hooks import (
     BackendUnavailable,
-    Pep517HookCaller,
+    BuildBackendHookCaller,
     UnsupportedOperation,
     default_subprocess_runner,
 )
@@ -26,7 +26,7 @@ def get_hooks(pkg, **kwargs):
     source_dir = pjoin(SAMPLES_DIR, pkg)
     with open(pjoin(source_dir, 'pyproject.toml'), 'rb') as f:
         data = tomllib.load(f)
-    return Pep517HookCaller(
+    return BuildBackendHookCaller(
         source_dir, data['build-system']['build-backend'], **kwargs
     )